filteropties toegevoegd

+++ File_system.py
@@ -1,25 +1,30 @@
 import os
 from datetime import datetime
 import tabulate
+from ast import literal_eval
 
 allfilesystem_list = []
 
-#om de list naar het bestand toe te schrijven en naar scherm te printen
-def savelist(listname):
-
-    #print tabel naar scherm
-    header = allfilesystem_list[0].keys()
-    rows = [x.values() for x in allfilesystem_list]
+# print tabel naar scherm
+def show_list(listname):
+    header = listname[0].keys()
+    rows = [x.values() for x in listname]
     tablefilesystem = tabulate.tabulate(rows, header, tablefmt='rst')
     print(tablefilesystem)
+    return tablefilesystem
 
-    #schrijf tabel weg naar bestand
-    f = open('C:\\Users\lucil\OneDrive\Documenten\school\Filesystem.txt', 'w')
-    f.write(tablefilesystem)
-    f.close()
+# schrijf tabel weg naar bestand
+def save_list(listname):
+    save = input('Do you want to save te results to a file? (Y/N)?: ')
+    if save == 'Y':
+        f = open('C:\\Users\lucil\OneDrive\Documenten\school\Filesystem.txt', 'w')
+        f.write(listname)
+        f.close()
 
-    #locatie moet nog worden bepaald
-    print('\nThe analysis is saved into a file called Filesystem.txt on the location ...')
+        # locatie moet nog worden bepaald
+        print('\nThe results are saved into a file called Filesystem.txt on the location ...')
+    else:
+        print('The results are not saved.')
 
 #om filesystem langs te lopen
 def analysefilesystem():
@@ -46,10 +51,50 @@ def analysefilesystem():
             file_size = file_stats.st_size
 
             allfilesystem_list.append({'File path': path, 'File name': f, 'File size (bytes)': file_size, 'last access': last_access, 'Last modified': last_modified, 'Creation time': creation_time})
+    return allfilesystem_list
+
+def filterfiles(listname):
+    filteredlist = []
+
+    #filteroptie op File size
+    size1 = input('Do you want to filter on file size? Y/N?: ')
+    if size1 == 'Y':
+        size2 = input('Type the file size you want to filter on. (bytes): ')
+        size2 = int(size2)
+        sizelist = filter(lambda x: x['File size (bytes)'] == size2, listname)
+        sizelist2 = list(sizelist)
+        sizestr = str(sizelist2)
+        sizestr2 = sizestr.replace('[', '')
+        sizestr3 = sizestr2.replace(']', '')
+        sizelistdict = literal_eval(sizestr3)
+        filteredlist.append(sizelistdict)
+    else:
+        print('ok')
+
+    #filteroptie op File name
+    name1 = input('Do you want to filter on file name? Y/N?: ')
+    if name1 == 'Y':
+        name2 = input('Type the name of the file you want to filter on: ')
+        namelist = filter(lambda x: x['File name'] == name2, listname)
+        namelist2 = list(namelist)
+        namestr = str(namelist2)
+        namestr2 = namestr.replace('[', '')
+        namestr3 = namestr2.replace(']', '')
+        namelistdict = literal_eval(namestr3)
+        filteredlist.append(namelistdict)
+    else:
+        print('ok')
+
+    return filteredlist
 
 
 if __name__ == '__main__':
 
-    analysefilesystem()
-    savelist(allfilesystem_list)
+    resultlist = analysefilesystem()
+    tablelist = show_list(resultlist)
+    filtervraag = input('Do you want to filter the results? (Y/N)?: ')
+    if filtervraag == 'Y':
+        filteredlist = filterfiles(allfilesystem_list)
+        tablelist = show_list(filteredlist)
+    save_list(tablelist)
